refactor(rest-and-cancellations): decompose god module into façade + 4 sub-modules
Extract 8 functions from rest_and_cancellations.py (1013 LOC) into
workflows/rest/ package (veto_check, planning_ops, markdown_entries,
reconciliation). Façade retains orchestrator + re-exports for backward
compatibility. 7th god script refactored in Phase 3.
