+++ magma_cycling/workflows/rest/__init__.py
@@ -1,0 +1,1 @@
+"""Rest and cancellations sub-modules for god module decomposition."""
